refactor: organise project structure and make code more consistent
Update project structure to include public/, src/main/, src/renderer/, and src/types/ directories to improve code organisation. Move relevant files to sensible locations. Update config files accordingly to make sure the app still runs.

Standardise naming convention of React component files to mirror the name of the component contained within the file.
E.g. Editor.tsx contains a component named Editor.

Standardise React functional component declaration style to be `export default function ...`.
