+++ src/renderer/components/Editor.tsx
@@ -1,12 +1,12 @@
 import { EditorProvider, FloatingMenu, BubbleMenu } from '@tiptap/react';
 import StarterKit from '@tiptap/starter-kit';
 import { Color } from '@tiptap/extension-color';
-import EditorJSONPreview from './jsonPreview';
-import MyEditorToolbar from './editorToolbar';
+import JSONPreview from './JSONPreview';
+import EditorToolbar from './EditorToolbar';
 import Underline from '@tiptap/extension-underline';
 import TextStyle from '@tiptap/extension-text-style';
 import CharacterCount from '@tiptap/extension-character-count';
-import WordCounter from './wordCounter';
+import WordCounter from './WordCounter';
 import CodeBlock from '@tiptap/extension-code-block';
 import Focus from '@tiptap/extension-focus';
 
@@ -14,7 +14,7 @@ import Highlight from '@tiptap/extension-highlight';
 import TextAlign from '@tiptap/extension-text-align';
 import Typography from '@tiptap/extension-typography';
 
-// define your extension array
+// Define extension array
 const extensions = [
   StarterKit,
   Highlight,
@@ -39,21 +39,19 @@ const content = '<p>Begin typing...</p>';
 const editorFooter = (
   <>
     <WordCounter />
-    <EditorJSONPreview />
+    <JSONPreview />
   </>
 );
 
-const Tiptap = () => {
+export default function Editor() {
   return (
     <>
       <EditorProvider
         extensions={extensions}
         content={content}
-        slotBefore={<MyEditorToolbar />}
+        slotBefore={<EditorToolbar />}
         slotAfter={editorFooter}
       ></EditorProvider>
     </>
   );
-};
-
-export default Tiptap;
+}
