+++ src/renderer/components/WordCounter.tsx
