+++ src/renderer/renderer.ts
@@ -26,8 +26,8 @@
  * ```
  */
 
-import './app';
-import './index.css';
+import './App';
+import '../../public/index.css';
 
 console.log(
   '👋 This message is being logged by "renderer.js", included via webpack'
