+++ webpack.main.config.ts
@@ -8,7 +8,7 @@ export const mainConfig: Configuration = {
    * This is the main entry point for your application, it's the first file
    * that runs in the main process.
    */
-  entry: './src/index.tsx',
+  entry: './src/renderer/index.tsx',
   // Put your normal webpack config below here
   module: {
     rules,
